[Dev] Subiendo Ticket Dto - mapper

+++ src/main/java/com/tour/events/domain/repository/EventDtoRepository.java
@@ -7,6 +7,6 @@ import java.util.Optional;
 
 public interface EventDtoRepository {
     List<EventDto> getAll();
-    Optional<EventDto> getById(long idEventDto);
+    Optional<EventDto> getById(Integer idEventDto);
     EventDto save(EventDto ticketDto);
 }
