+++ src/main/java/com/tour/events/infraestructure/repositories/EventCrudRepository.java
@@ -3,5 +3,5 @@ package com.tour.events.infraestructure.repositories;
 import com.tour.events.infraestructure.entities.Event;
 import org.springframework.data.repository.CrudRepository;
 
-public interface EventCrudRepository extends CrudRepository<Event, Long> {
+public interface EventCrudRepository extends CrudRepository<Event, Integer> {
 }
