+++ src/main/java/com/tour/events/infraestructure/repositories/TicketCrudRepository.java
@@ -3,5 +3,5 @@ package com.tour.events.infraestructure.repositories;
 import com.tour.events.infraestructure.entities.Ticket;
 import org.springframework.data.repository.CrudRepository;
 
-public interface TicketCrudRepository  extends CrudRepository<Ticket, Long> {
+public interface TicketCrudRepository  extends CrudRepository<Ticket, Integer> {
 }
